Small Fix
Small fix in PresentCommands.java - if a players TP wasn't set it sent
the command sender two messages. The 'forgot to set TP' message and
'player (player) hasn't asked for a gift) message. This was due to a
miss placed '}' placed before the [remove player from "Names" list].
This also fixed the players being removed from the "Names" list when
doing the done command.